refactor: move queries.ts from runtime/ to core/
