+++ packages/db/src/core/integration/index.ts
@@ -6,7 +6,7 @@ import { mkdir, writeFile } from 'fs/promises';
 import { blue, yellow } from 'kleur/colors';
 import { loadEnv } from 'vite';
 import parseArgs from 'yargs-parser';
-import { SEED_DEV_FILE_NAME } from '../../runtime/queries.js';
+import { SEED_DEV_FILE_NAME } from '../queries.js';
 import { AstroDbError } from '../../runtime/utils.js';
 import { CONFIG_FILE_NAMES, DB_PATH } from '../consts.js';
 import { resolveDbConfig } from '../load-file.js';
